cleaning up devices, moving to using generic classes and using special ones only when nessesary
