--- rexus/config/Display.py
@@ -1,5 +1,0 @@
-from Phidgets.Devices.TextLCD import TextLCD_ScreenSize
-
-BACKLIGHT = True
-CONTRAST = 180
-SIZE = TextLCD_ScreenSize.PHIDGET_TEXTLCD_SCREEN_4x40
